@@ -1,4 +1,4 @@
-const list = require('./list')
+import list from './list.js'
 
 const diff = async ({source, target}) => {
   const left = await list(source)
@@ -11,4 +11,4 @@ const diff = async ({source, target}) => {
     .map(file => file.name)
 }
 
-module.exports = diff
+export default diff
