@@ -1,5 +1,5 @@
-const test = require('tape')
-const diff = require('./diff')
+import test from 'tape'
+import diff from './diff.js'
 
 test('diff files, when using same tree', async assert => {
   const message = 'diff result must be empty'
