@@ -1,5 +1,5 @@
-const path = require('path')
-const fs = require('fs').promises
+import path from 'node:path'
+import fs from 'node:fs/promises'
 
 const list = async pathdir => {
   try {
@@ -27,4 +27,4 @@ const list = async pathdir => {
   }
 }
 
-module.exports = list
+export default list
