@@ -1,5 +1,5 @@
-const test = require('tape')
-const list = require('./list')
+import test from 'tape'
+import list from './list.js'
 
 test('list files, when using flat tree', async assert => {
   const message = 'should list 10 files'
